Revert "MergedDescriptorValidator"
This reverts commit 4b8d64ee308b8fcbc4e2369da12c84d99d301863.
